@@ -296,21 +296,22 @@ func RemoveRoleDS(ms *dstate.MemberState, role int64) error {
 }
 
 var StringPerms = map[int64]string{
-	discordgo.PermissionViewChannel:        "View Channel",
-	discordgo.PermissionSendMessages:       "Send Messages",
-	discordgo.PermissionSendTTSMessages:    "Send TTS Messages",
-	discordgo.PermissionManageMessages:     "Manage Messages",
-	discordgo.PermissionEmbedLinks:         "Embed Links",
-	discordgo.PermissionAttachFiles:        "Attach Files",
-	discordgo.PermissionReadMessageHistory: "Read Message History",
-	discordgo.PermissionMentionEveryone:    "Mention Everyone",
-	discordgo.PermissionVoiceConnect:       "Voice Connect",
-	discordgo.PermissionVoiceSpeak:         "Voice Speak",
-	discordgo.PermissionVoiceMuteMembers:   "Voice Mute Members",
-	discordgo.PermissionVoiceDeafenMembers: "Voice Deafen Members",
-	discordgo.PermissionVoiceMoveMembers:   "Voice Move Members",
-	discordgo.PermissionVoiceUseVAD:        "Voice Use VAD",
-
+	discordgo.PermissionViewChannel:         "View Channel",
+	discordgo.PermissionSendMessages:        "Send Messages",
+	discordgo.PermissionSendTTSMessages:     "Send TTS Messages",
+	discordgo.PermissionManageMessages:      "Manage Messages",
+	discordgo.PermissionEmbedLinks:          "Embed Links",
+	discordgo.PermissionAttachFiles:         "Attach Files",
+	discordgo.PermissionReadMessageHistory:  "Read Message History",
+	discordgo.PermissionMentionEveryone:     "Mention Everyone",
+	discordgo.PermissionVoiceConnect:        "Voice Connect",
+	discordgo.PermissionVoiceSpeak:          "Voice Speak",
+	discordgo.PermissionVoiceMuteMembers:    "Voice Mute Members",
+	discordgo.PermissionVoiceDeafenMembers:  "Voice Deafen Members",
+	discordgo.PermissionVoiceMoveMembers:    "Voice Move Members",
+	discordgo.PermissionVoiceUseVAD:         "Voice Use VAD",
+	discordgo.PermissionChangeNickname:      "Change Nickname",
+	discordgo.PermissionManageNicknames:     "Manage Nicknames",
 	discordgo.PermissionCreateInstantInvite: "Create Instant Invite",
 	discordgo.PermissionKickMembers:         "Kick Members",
 	discordgo.PermissionBanMembers:          "Ban Members",
