@@ -130,6 +130,7 @@ func (p *Plugin) InitWeb() {
 	subMux.Handle(pat.Post("/database/delete/:id"), web.ControllerPostHandler(handleDeleteDatabaseEntry, getDBHandler, nil))
 
 	subMux.Handle(pat.Get("/commands/:cmd/"), getCmdHandler)
+	subMux.Handle(pat.Get("/commands/:cmd"), getCmdHandler)
 
 	subMux.Handle(pat.Get("/groups/:group/"), web.ControllerHandler(handleGetCommandsGroup, "cp_custom_commands"))
 	subMux.Handle(pat.Get("/groups/:group"), web.ControllerHandler(handleGetCommandsGroup, "cp_custom_commands"))
@@ -421,6 +422,8 @@ func handleNewCommand(w http.ResponseWriter, r *http.Request) (web.TemplateData,
 		return templateData, errors.WrapIf(err, "error generating local id")
 	}
 
+	defaultResponse := fmt.Sprintf("Response generated by Custom Command {{.CCID}}. This response can be edited using [the bot's control panel.](<%s/manage/{{.Guild.ID}}/customcommands/commands/{{.CCID}}>)", web.BaseURL())
+
 	dbModel := &models.CustomCommand{
 		GuildID: activeGuild.ID,
 		LocalID: localID,
@@ -431,7 +434,7 @@ func handleNewCommand(w http.ResponseWriter, r *http.Request) (web.TemplateData,
 		TimeTriggerExcludingDays:  []int64{},
 		TimeTriggerExcludingHours: []int64{},
 
-		Responses: []string{"Edit this to change the output of the custom command {{.CCID}}!"},
+		Responses: []string{defaultResponse},
 	}
 
 	if groupID != 0 {
