@@ -19,9 +19,9 @@ func TestEncode(t *testing.T) {
 		numFrames++
 	}
 
-	// Predermined, probably gonna change the testing method somehow
-	if numFrames != 756 {
-		t.Errorf("Incorrect number of frames (got %d expected %d)", numFrames, 756)
+	// Predetermined, probably gonna change the testing method somehow
+	if numFrames != 755 {
+		t.Errorf("Incorrect number of frames (got %d expected %d)", numFrames, 755)
 		t.Fail()
 	}
 }
