@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
