@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -35,6 +35,7 @@ type GuildLoggingConfig struct {
 	EveryoneCanViewDeleted       null.Bool        `boil:"everyone_can_view_deleted" json:"everyone_can_view_deleted,omitempty" toml:"everyone_can_view_deleted" yaml:"everyone_can_view_deleted,omitempty"`
 	MessageLogsAllowedRoles      types.Int64Array `boil:"message_logs_allowed_roles" json:"message_logs_allowed_roles,omitempty" toml:"message_logs_allowed_roles" yaml:"message_logs_allowed_roles,omitempty"`
 	AccessMode                   int16            `boil:"access_mode" json:"access_mode" toml:"access_mode" yaml:"access_mode"`
+	ChannelsWhitelistMode        bool             `boil:"channels_whitelist_mode" json:"channels_whitelist_mode" toml:"channels_whitelist_mode" yaml:"channels_whitelist_mode"`
 
 	R *guildLoggingConfigR `boil:"-" json:"-" toml:"-" yaml:"-"`
 	L guildLoggingConfigL  `boil:"-" json:"-" toml:"-" yaml:"-"`
@@ -51,6 +52,7 @@ var GuildLoggingConfigColumns = struct {
 	EveryoneCanViewDeleted       string
 	MessageLogsAllowedRoles      string
 	AccessMode                   string
+	ChannelsWhitelistMode        string
 }{
 	GuildID:                      "guild_id",
 	CreatedAt:                    "created_at",
@@ -62,6 +64,33 @@ var GuildLoggingConfigColumns = struct {
 	EveryoneCanViewDeleted:       "everyone_can_view_deleted",
 	MessageLogsAllowedRoles:      "message_logs_allowed_roles",
 	AccessMode:                   "access_mode",
+	ChannelsWhitelistMode:        "channels_whitelist_mode",
+}
+
+var GuildLoggingConfigTableColumns = struct {
+	GuildID                      string
+	CreatedAt                    string
+	UpdatedAt                    string
+	UsernameLoggingEnabled       string
+	NicknameLoggingEnabled       string
+	BlacklistedChannels          string
+	ManageMessagesCanViewDeleted string
+	EveryoneCanViewDeleted       string
+	MessageLogsAllowedRoles      string
+	AccessMode                   string
+	ChannelsWhitelistMode        string
+}{
+	GuildID:                      "guild_logging_configs.guild_id",
+	CreatedAt:                    "guild_logging_configs.created_at",
+	UpdatedAt:                    "guild_logging_configs.updated_at",
+	UsernameLoggingEnabled:       "guild_logging_configs.username_logging_enabled",
+	NicknameLoggingEnabled:       "guild_logging_configs.nickname_logging_enabled",
+	BlacklistedChannels:          "guild_logging_configs.blacklisted_channels",
+	ManageMessagesCanViewDeleted: "guild_logging_configs.manage_messages_can_view_deleted",
+	EveryoneCanViewDeleted:       "guild_logging_configs.everyone_can_view_deleted",
+	MessageLogsAllowedRoles:      "guild_logging_configs.message_logs_allowed_roles",
+	AccessMode:                   "guild_logging_configs.access_mode",
+	ChannelsWhitelistMode:        "guild_logging_configs.channels_whitelist_mode",
 }
 
 // Generated where
@@ -97,8 +126,6 @@ func (w whereHelpernull_Time) EQ(x null.Time) qm.QueryMod {
 func (w whereHelpernull_Time) NEQ(x null.Time) qm.QueryMod {
 	return qmhelper.WhereNullEQ(w.field, true, x)
 }
-func (w whereHelpernull_Time) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
-func (w whereHelpernull_Time) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 func (w whereHelpernull_Time) LT(x null.Time) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.LT, x)
 }
@@ -112,6 +139,9 @@ func (w whereHelpernull_Time) GTE(x null.Time) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
 
+func (w whereHelpernull_Time) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
+func (w whereHelpernull_Time) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
+
 type whereHelpernull_Bool struct{ field string }
 
 func (w whereHelpernull_Bool) EQ(x null.Bool) qm.QueryMod {
@@ -120,8 +150,6 @@ func (w whereHelpernull_Bool) EQ(x null.Bool) qm.QueryMod {
 func (w whereHelpernull_Bool) NEQ(x null.Bool) qm.QueryMod {
 	return qmhelper.WhereNullEQ(w.field, true, x)
 }
-func (w whereHelpernull_Bool) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
-func (w whereHelpernull_Bool) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 func (w whereHelpernull_Bool) LT(x null.Bool) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.LT, x)
 }
@@ -135,6 +163,9 @@ func (w whereHelpernull_Bool) GTE(x null.Bool) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
 
+func (w whereHelpernull_Bool) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
+func (w whereHelpernull_Bool) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
+
 type whereHelpernull_String struct{ field string }
 
 func (w whereHelpernull_String) EQ(x null.String) qm.QueryMod {
@@ -143,8 +174,6 @@ func (w whereHelpernull_String) EQ(x null.String) qm.QueryMod {
 func (w whereHelpernull_String) NEQ(x null.String) qm.QueryMod {
 	return qmhelper.WhereNullEQ(w.field, true, x)
 }
-func (w whereHelpernull_String) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
-func (w whereHelpernull_String) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 func (w whereHelpernull_String) LT(x null.String) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.LT, x)
 }
@@ -157,6 +186,23 @@ func (w whereHelpernull_String) GT(x null.String) qm.QueryMod {
 func (w whereHelpernull_String) GTE(x null.String) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
+func (w whereHelpernull_String) IN(slice []string) qm.QueryMod {
+	values := make([]interface{}, 0, len(slice))
+	for _, value := range slice {
+		values = append(values, value)
+	}
+	return qm.WhereIn(fmt.Sprintf("%s IN ?", w.field), values...)
+}
+func (w whereHelpernull_String) NIN(slice []string) qm.QueryMod {
+	values := make([]interface{}, 0, len(slice))
+	for _, value := range slice {
+		values = append(values, value)
+	}
+	return qm.WhereNotIn(fmt.Sprintf("%s NOT IN ?", w.field), values...)
+}
+
+func (w whereHelpernull_String) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
+func (w whereHelpernull_String) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 
 type whereHelpertypes_Int64Array struct{ field string }
 
@@ -166,8 +212,6 @@ func (w whereHelpertypes_Int64Array) EQ(x types.Int64Array) qm.QueryMod {
 func (w whereHelpertypes_Int64Array) NEQ(x types.Int64Array) qm.QueryMod {
 	return qmhelper.WhereNullEQ(w.field, true, x)
 }
-func (w whereHelpertypes_Int64Array) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
-func (w whereHelpertypes_Int64Array) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 func (w whereHelpertypes_Int64Array) LT(x types.Int64Array) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.LT, x)
 }
@@ -180,6 +224,9 @@ func (w whereHelpertypes_Int64Array) GT(x types.Int64Array) qm.QueryMod {
 func (w whereHelpertypes_Int64Array) GTE(x types.Int64Array) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
+
+func (w whereHelpertypes_Int64Array) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
+func (w whereHelpertypes_Int64Array) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 
 type whereHelperint16 struct{ field string }
 
@@ -204,6 +251,15 @@ func (w whereHelperint16) NIN(slice []int16) qm.QueryMod {
 	return qm.WhereNotIn(fmt.Sprintf("%s NOT IN ?", w.field), values...)
 }
 
+type whereHelperbool struct{ field string }
+
+func (w whereHelperbool) EQ(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.EQ, x) }
+func (w whereHelperbool) NEQ(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.NEQ, x) }
+func (w whereHelperbool) LT(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.LT, x) }
+func (w whereHelperbool) LTE(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.LTE, x) }
+func (w whereHelperbool) GT(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.GT, x) }
+func (w whereHelperbool) GTE(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.GTE, x) }
+
 var GuildLoggingConfigWhere = struct {
 	GuildID                      whereHelperint64
 	CreatedAt                    whereHelpernull_Time
@@ -215,6 +271,7 @@ var GuildLoggingConfigWhere = struct {
 	EveryoneCanViewDeleted       whereHelpernull_Bool
 	MessageLogsAllowedRoles      whereHelpertypes_Int64Array
 	AccessMode                   whereHelperint16
+	ChannelsWhitelistMode        whereHelperbool
 }{
 	GuildID:                      whereHelperint64{field: "\"guild_logging_configs\".\"guild_id\""},
 	CreatedAt:                    whereHelpernull_Time{field: "\"guild_logging_configs\".\"created_at\""},
@@ -226,6 +283,7 @@ var GuildLoggingConfigWhere = struct {
 	EveryoneCanViewDeleted:       whereHelpernull_Bool{field: "\"guild_logging_configs\".\"everyone_can_view_deleted\""},
 	MessageLogsAllowedRoles:      whereHelpertypes_Int64Array{field: "\"guild_logging_configs\".\"message_logs_allowed_roles\""},
 	AccessMode:                   whereHelperint16{field: "\"guild_logging_configs\".\"access_mode\""},
+	ChannelsWhitelistMode:        whereHelperbool{field: "\"guild_logging_configs\".\"channels_whitelist_mode\""},
 }
 
 // GuildLoggingConfigRels is where relationship names are stored.
@@ -245,15 +303,16 @@ func (*guildLoggingConfigR) NewStruct() *guildLoggingConfigR {
 type guildLoggingConfigL struct{}
 
 var (
-	guildLoggingConfigAllColumns            = []string{"guild_id", "created_at", "updated_at", "username_logging_enabled", "nickname_logging_enabled", "blacklisted_channels", "manage_messages_can_view_deleted", "everyone_can_view_deleted", "message_logs_allowed_roles", "access_mode"}
-	guildLoggingConfigColumnsWithoutDefault = []string{"created_at", "updated_at", "username_logging_enabled", "nickname_logging_enabled", "blacklisted_channels", "manage_messages_can_view_deleted", "everyone_can_view_deleted", "message_logs_allowed_roles"}
-	guildLoggingConfigColumnsWithDefault    = []string{"guild_id", "access_mode"}
+	guildLoggingConfigAllColumns            = []string{"guild_id", "created_at", "updated_at", "username_logging_enabled", "nickname_logging_enabled", "blacklisted_channels", "manage_messages_can_view_deleted", "everyone_can_view_deleted", "message_logs_allowed_roles", "access_mode", "channels_whitelist_mode"}
+	guildLoggingConfigColumnsWithoutDefault = []string{"guild_id"}
+	guildLoggingConfigColumnsWithDefault    = []string{"created_at", "updated_at", "username_logging_enabled", "nickname_logging_enabled", "blacklisted_channels", "manage_messages_can_view_deleted", "everyone_can_view_deleted", "message_logs_allowed_roles", "access_mode", "channels_whitelist_mode"}
 	guildLoggingConfigPrimaryKeyColumns     = []string{"guild_id"}
+	guildLoggingConfigGeneratedColumns      = []string{}
 )
 
 type (
 	// GuildLoggingConfigSlice is an alias for a slice of pointers to GuildLoggingConfig.
-	// This should generally be used opposed to []GuildLoggingConfig.
+	// This should almost always be used instead of []GuildLoggingConfig.
 	GuildLoggingConfigSlice []*GuildLoggingConfig
 
 	guildLoggingConfigQuery struct {
@@ -295,7 +354,7 @@ func (q guildLoggingConfigQuery) One(ctx context.Context, exec boil.ContextExecu
 
 	err := q.Bind(ctx, exec, o)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: failed to execute a one query for guild_logging_configs")
@@ -321,7 +380,7 @@ func (q guildLoggingConfigQuery) All(ctx context.Context, exec boil.ContextExecu
 	return o, nil
 }
 
-// CountG returns the count of all GuildLoggingConfig records in the query, and panics on error.
+// CountG returns the count of all GuildLoggingConfig records in the query using the global executor
 func (q guildLoggingConfigQuery) CountG(ctx context.Context) (int64, error) {
 	return q.Count(ctx, boil.GetContextDB())
 }
@@ -341,7 +400,7 @@ func (q guildLoggingConfigQuery) Count(ctx context.Context, exec boil.ContextExe
 	return count, nil
 }
 
-// ExistsG checks if the row exists in the table, and panics on error.
+// ExistsG checks if the row exists in the table using the global executor.
 func (q guildLoggingConfigQuery) ExistsG(ctx context.Context) (bool, error) {
 	return q.Exists(ctx, boil.GetContextDB())
 }
@@ -365,7 +424,12 @@ func (q guildLoggingConfigQuery) Exists(ctx context.Context, exec boil.ContextEx
 // GuildLoggingConfigs retrieves all the records using an executor.
 func GuildLoggingConfigs(mods ...qm.QueryMod) guildLoggingConfigQuery {
 	mods = append(mods, qm.From("\"guild_logging_configs\""))
-	return guildLoggingConfigQuery{NewQuery(mods...)}
+	q := NewQuery(mods...)
+	if len(queries.GetSelect(q)) == 0 {
+		queries.SetSelect(q, []string{"\"guild_logging_configs\".*"})
+	}
+
+	return guildLoggingConfigQuery{q}
 }
 
 // FindGuildLoggingConfigG retrieves a single record by ID.
@@ -390,7 +454,7 @@ func FindGuildLoggingConfig(ctx context.Context, exec boil.ContextExecutor, guil
 
 	err := q.Bind(ctx, exec, guildLoggingConfigObj)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: unable to select from guild_logging_configs")
@@ -637,13 +701,13 @@ func (o GuildLoggingConfigSlice) UpdateAll(ctx context.Context, exec boil.Contex
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *GuildLoggingConfig) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
+func (o *GuildLoggingConfig) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *GuildLoggingConfig) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+func (o *GuildLoggingConfig) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
 	if o == nil {
 		return errors.New("models: no guild_logging_configs provided for upsert")
 	}
@@ -693,12 +757,13 @@ func (o *GuildLoggingConfig) Upsert(ctx context.Context, exec boil.ContextExecut
 	var err error
 
 	if !cached {
-		insert, ret := insertColumns.InsertColumnSet(
+		insert, _ := insertColumns.InsertColumnSet(
 			guildLoggingConfigAllColumns,
 			guildLoggingConfigColumnsWithDefault,
 			guildLoggingConfigColumnsWithoutDefault,
 			nzDefaults,
 		)
+
 		update := updateColumns.UpdateColumnSet(
 			guildLoggingConfigAllColumns,
 			guildLoggingConfigPrimaryKeyColumns,
@@ -708,12 +773,18 @@ func (o *GuildLoggingConfig) Upsert(ctx context.Context, exec boil.ContextExecut
 			return errors.New("models: unable to upsert guild_logging_configs, could not build update column list")
 		}
 
+		ret := strmangle.SetComplement(guildLoggingConfigAllColumns, strmangle.SetIntersect(insert, update))
+
 		conflict := conflictColumns
-		if len(conflict) == 0 {
+		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
+			if len(guildLoggingConfigPrimaryKeyColumns) == 0 {
+				return errors.New("models: unable to upsert guild_logging_configs, could not build conflict column list")
+			}
+
 			conflict = make([]string, len(guildLoggingConfigPrimaryKeyColumns))
 			copy(conflict, guildLoggingConfigPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"guild_logging_configs\"", updateOnConflict, ret, update, conflict, insert)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"guild_logging_configs\"", updateOnConflict, ret, update, conflict, insert, opts...)
 
 		cache.valueMapping, err = queries.BindMapping(guildLoggingConfigType, guildLoggingConfigMapping, insert)
 		if err != nil {
@@ -741,7 +812,7 @@ func (o *GuildLoggingConfig) Upsert(ctx context.Context, exec boil.ContextExecut
 	}
 	if len(cache.retMapping) != 0 {
 		err = exec.QueryRowContext(ctx, cache.query, vals...).Scan(returns...)
-		if err == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			err = nil // Postgres doesn't return anything when there's no update
 		}
 	} else {
@@ -940,4 +1011,9 @@ func GuildLoggingConfigExists(ctx context.Context, exec boil.ContextExecutor, gu
 	}
 
 	return exists, nil
+}
+
+// Exists checks if the GuildLoggingConfig row exists.
+func (o *GuildLoggingConfig) Exists(ctx context.Context, exec boil.ContextExecutor) (bool, error) {
+	return GuildLoggingConfigExists(ctx, exec, o.GuildID)
 }
