@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -61,6 +61,30 @@ var MessageLogs2Columns = struct {
 	AuthorID:       "author_id",
 	AuthorUsername: "author_username",
 	Messages:       "messages",
+}
+
+var MessageLogs2TableColumns = struct {
+	ID             string
+	GuildID        string
+	LegacyID       string
+	CreatedAt      string
+	UpdatedAt      string
+	ChannelName    string
+	ChannelID      string
+	AuthorID       string
+	AuthorUsername string
+	Messages       string
+}{
+	ID:             "message_logs2.id",
+	GuildID:        "message_logs2.guild_id",
+	LegacyID:       "message_logs2.legacy_id",
+	CreatedAt:      "message_logs2.created_at",
+	UpdatedAt:      "message_logs2.updated_at",
+	ChannelName:    "message_logs2.channel_name",
+	ChannelID:      "message_logs2.channel_id",
+	AuthorID:       "message_logs2.author_id",
+	AuthorUsername: "message_logs2.author_username",
+	Messages:       "message_logs2.messages",
 }
 
 // Generated where
@@ -174,14 +198,15 @@ type messageLogs2L struct{}
 
 var (
 	messageLogs2AllColumns            = []string{"id", "guild_id", "legacy_id", "created_at", "updated_at", "channel_name", "channel_id", "author_id", "author_username", "messages"}
-	messageLogs2ColumnsWithoutDefault = []string{"id", "guild_id", "legacy_id", "created_at", "updated_at", "channel_name", "channel_id", "author_id", "author_username", "messages"}
-	messageLogs2ColumnsWithDefault    = []string{}
+	messageLogs2ColumnsWithoutDefault = []string{"id", "guild_id", "legacy_id", "created_at", "updated_at", "channel_name", "channel_id", "author_id", "author_username"}
+	messageLogs2ColumnsWithDefault    = []string{"messages"}
 	messageLogs2PrimaryKeyColumns     = []string{"guild_id", "id"}
+	messageLogs2GeneratedColumns      = []string{}
 )
 
 type (
 	// MessageLogs2Slice is an alias for a slice of pointers to MessageLogs2.
-	// This should generally be used opposed to []MessageLogs2.
+	// This should almost always be used instead of []MessageLogs2.
 	MessageLogs2Slice []*MessageLogs2
 
 	messageLogs2Query struct {
@@ -223,7 +248,7 @@ func (q messageLogs2Query) One(ctx context.Context, exec boil.ContextExecutor) (
 
 	err := q.Bind(ctx, exec, o)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: failed to execute a one query for message_logs2")
@@ -249,7 +274,7 @@ func (q messageLogs2Query) All(ctx context.Context, exec boil.ContextExecutor) (
 	return o, nil
 }
 
-// CountG returns the count of all MessageLogs2 records in the query, and panics on error.
+// CountG returns the count of all MessageLogs2 records in the query using the global executor
 func (q messageLogs2Query) CountG(ctx context.Context) (int64, error) {
 	return q.Count(ctx, boil.GetContextDB())
 }
@@ -269,7 +294,7 @@ func (q messageLogs2Query) Count(ctx context.Context, exec boil.ContextExecutor)
 	return count, nil
 }
 
-// ExistsG checks if the row exists in the table, and panics on error.
+// ExistsG checks if the row exists in the table using the global executor.
 func (q messageLogs2Query) ExistsG(ctx context.Context) (bool, error) {
 	return q.Exists(ctx, boil.GetContextDB())
 }
@@ -293,7 +318,12 @@ func (q messageLogs2Query) Exists(ctx context.Context, exec boil.ContextExecutor
 // MessageLogs2s retrieves all the records using an executor.
 func MessageLogs2s(mods ...qm.QueryMod) messageLogs2Query {
 	mods = append(mods, qm.From("\"message_logs2\""))
-	return messageLogs2Query{NewQuery(mods...)}
+	q := NewQuery(mods...)
+	if len(queries.GetSelect(q)) == 0 {
+		queries.SetSelect(q, []string{"\"message_logs2\".*"})
+	}
+
+	return messageLogs2Query{q}
 }
 
 // FindMessageLogs2G retrieves a single record by ID.
@@ -318,7 +348,7 @@ func FindMessageLogs2(ctx context.Context, exec boil.ContextExecutor, guildID in
 
 	err := q.Bind(ctx, exec, messageLogs2Obj)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: unable to select from message_logs2")
@@ -565,13 +595,13 @@ func (o MessageLogs2Slice) UpdateAll(ctx context.Context, exec boil.ContextExecu
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *MessageLogs2) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
+func (o *MessageLogs2) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *MessageLogs2) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+func (o *MessageLogs2) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
 	if o == nil {
 		return errors.New("models: no message_logs2 provided for upsert")
 	}
@@ -621,12 +651,13 @@ func (o *MessageLogs2) Upsert(ctx context.Context, exec boil.ContextExecutor, up
 	var err error
 
 	if !cached {
-		insert, ret := insertColumns.InsertColumnSet(
+		insert, _ := insertColumns.InsertColumnSet(
 			messageLogs2AllColumns,
 			messageLogs2ColumnsWithDefault,
 			messageLogs2ColumnsWithoutDefault,
 			nzDefaults,
 		)
+
 		update := updateColumns.UpdateColumnSet(
 			messageLogs2AllColumns,
 			messageLogs2PrimaryKeyColumns,
@@ -636,12 +667,18 @@ func (o *MessageLogs2) Upsert(ctx context.Context, exec boil.ContextExecutor, up
 			return errors.New("models: unable to upsert message_logs2, could not build update column list")
 		}
 
+		ret := strmangle.SetComplement(messageLogs2AllColumns, strmangle.SetIntersect(insert, update))
+
 		conflict := conflictColumns
-		if len(conflict) == 0 {
+		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
+			if len(messageLogs2PrimaryKeyColumns) == 0 {
+				return errors.New("models: unable to upsert message_logs2, could not build conflict column list")
+			}
+
 			conflict = make([]string, len(messageLogs2PrimaryKeyColumns))
 			copy(conflict, messageLogs2PrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"message_logs2\"", updateOnConflict, ret, update, conflict, insert)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"message_logs2\"", updateOnConflict, ret, update, conflict, insert, opts...)
 
 		cache.valueMapping, err = queries.BindMapping(messageLogs2Type, messageLogs2Mapping, insert)
 		if err != nil {
@@ -669,7 +706,7 @@ func (o *MessageLogs2) Upsert(ctx context.Context, exec boil.ContextExecutor, up
 	}
 	if len(cache.retMapping) != 0 {
 		err = exec.QueryRowContext(ctx, cache.query, vals...).Scan(returns...)
-		if err == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			err = nil // Postgres doesn't return anything when there's no update
 		}
 	} else {
@@ -868,4 +905,9 @@ func MessageLogs2Exists(ctx context.Context, exec boil.ContextExecutor, guildID 
 	}
 
 	return exists, nil
+}
+
+// Exists checks if the MessageLogs2 row exists.
+func (o *MessageLogs2) Exists(ctx context.Context, exec boil.ContextExecutor) (bool, error) {
+	return MessageLogs2Exists(ctx, exec, o.GuildID, o.ID)
 }
