@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -56,16 +56,27 @@ var Messages2Columns = struct {
 	Content:        "content",
 }
 
+var Messages2TableColumns = struct {
+	ID             string
+	GuildID        string
+	CreatedAt      string
+	UpdatedAt      string
+	Deleted        string
+	AuthorUsername string
+	AuthorID       string
+	Content        string
+}{
+	ID:             "messages2.id",
+	GuildID:        "messages2.guild_id",
+	CreatedAt:      "messages2.created_at",
+	UpdatedAt:      "messages2.updated_at",
+	Deleted:        "messages2.deleted",
+	AuthorUsername: "messages2.author_username",
+	AuthorID:       "messages2.author_id",
+	Content:        "messages2.content",
+}
+
 // Generated where
-
-type whereHelperbool struct{ field string }
-
-func (w whereHelperbool) EQ(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.EQ, x) }
-func (w whereHelperbool) NEQ(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.NEQ, x) }
-func (w whereHelperbool) LT(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.LT, x) }
-func (w whereHelperbool) LTE(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.LTE, x) }
-func (w whereHelperbool) GT(x bool) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.GT, x) }
-func (w whereHelperbool) GTE(x bool) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.GTE, x) }
 
 var Messages2Where = struct {
 	ID             whereHelperint64
@@ -108,11 +119,12 @@ var (
 	messages2ColumnsWithoutDefault = []string{"id", "guild_id", "created_at", "updated_at", "deleted", "author_username", "author_id", "content"}
 	messages2ColumnsWithDefault    = []string{}
 	messages2PrimaryKeyColumns     = []string{"id"}
+	messages2GeneratedColumns      = []string{}
 )
 
 type (
 	// Messages2Slice is an alias for a slice of pointers to Messages2.
-	// This should generally be used opposed to []Messages2.
+	// This should almost always be used instead of []Messages2.
 	Messages2Slice []*Messages2
 
 	messages2Query struct {
@@ -154,7 +166,7 @@ func (q messages2Query) One(ctx context.Context, exec boil.ContextExecutor) (*Me
 
 	err := q.Bind(ctx, exec, o)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: failed to execute a one query for messages2")
@@ -180,7 +192,7 @@ func (q messages2Query) All(ctx context.Context, exec boil.ContextExecutor) (Mes
 	return o, nil
 }
 
-// CountG returns the count of all Messages2 records in the query, and panics on error.
+// CountG returns the count of all Messages2 records in the query using the global executor
 func (q messages2Query) CountG(ctx context.Context) (int64, error) {
 	return q.Count(ctx, boil.GetContextDB())
 }
@@ -200,7 +212,7 @@ func (q messages2Query) Count(ctx context.Context, exec boil.ContextExecutor) (i
 	return count, nil
 }
 
-// ExistsG checks if the row exists in the table, and panics on error.
+// ExistsG checks if the row exists in the table using the global executor.
 func (q messages2Query) ExistsG(ctx context.Context) (bool, error) {
 	return q.Exists(ctx, boil.GetContextDB())
 }
@@ -224,7 +236,12 @@ func (q messages2Query) Exists(ctx context.Context, exec boil.ContextExecutor) (
 // Messages2s retrieves all the records using an executor.
 func Messages2s(mods ...qm.QueryMod) messages2Query {
 	mods = append(mods, qm.From("\"messages2\""))
-	return messages2Query{NewQuery(mods...)}
+	q := NewQuery(mods...)
+	if len(queries.GetSelect(q)) == 0 {
+		queries.SetSelect(q, []string{"\"messages2\".*"})
+	}
+
+	return messages2Query{q}
 }
 
 // FindMessages2G retrieves a single record by ID.
@@ -249,7 +266,7 @@ func FindMessages2(ctx context.Context, exec boil.ContextExecutor, iD int64, sel
 
 	err := q.Bind(ctx, exec, messages2Obj)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: unable to select from messages2")
@@ -496,13 +513,13 @@ func (o Messages2Slice) UpdateAll(ctx context.Context, exec boil.ContextExecutor
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *Messages2) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
+func (o *Messages2) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *Messages2) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+func (o *Messages2) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
 	if o == nil {
 		return errors.New("models: no messages2 provided for upsert")
 	}
@@ -552,12 +569,13 @@ func (o *Messages2) Upsert(ctx context.Context, exec boil.ContextExecutor, updat
 	var err error
 
 	if !cached {
-		insert, ret := insertColumns.InsertColumnSet(
+		insert, _ := insertColumns.InsertColumnSet(
 			messages2AllColumns,
 			messages2ColumnsWithDefault,
 			messages2ColumnsWithoutDefault,
 			nzDefaults,
 		)
+
 		update := updateColumns.UpdateColumnSet(
 			messages2AllColumns,
 			messages2PrimaryKeyColumns,
@@ -567,12 +585,18 @@ func (o *Messages2) Upsert(ctx context.Context, exec boil.ContextExecutor, updat
 			return errors.New("models: unable to upsert messages2, could not build update column list")
 		}
 
+		ret := strmangle.SetComplement(messages2AllColumns, strmangle.SetIntersect(insert, update))
+
 		conflict := conflictColumns
-		if len(conflict) == 0 {
+		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
+			if len(messages2PrimaryKeyColumns) == 0 {
+				return errors.New("models: unable to upsert messages2, could not build conflict column list")
+			}
+
 			conflict = make([]string, len(messages2PrimaryKeyColumns))
 			copy(conflict, messages2PrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"messages2\"", updateOnConflict, ret, update, conflict, insert)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"messages2\"", updateOnConflict, ret, update, conflict, insert, opts...)
 
 		cache.valueMapping, err = queries.BindMapping(messages2Type, messages2Mapping, insert)
 		if err != nil {
@@ -600,7 +624,7 @@ func (o *Messages2) Upsert(ctx context.Context, exec boil.ContextExecutor, updat
 	}
 	if len(cache.retMapping) != 0 {
 		err = exec.QueryRowContext(ctx, cache.query, vals...).Scan(returns...)
-		if err == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			err = nil // Postgres doesn't return anything when there's no update
 		}
 	} else {
@@ -799,4 +823,9 @@ func Messages2Exists(ctx context.Context, exec boil.ContextExecutor, iD int64) (
 	}
 
 	return exists, nil
+}
+
+// Exists checks if the Messages2 row exists.
+func (o *Messages2) Exists(ctx context.Context, exec boil.ContextExecutor) (bool, error) {
+	return Messages2Exists(ctx, exec, o.ID)
 }
