@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -54,6 +54,24 @@ var NicknameListingColumns = struct {
 	Nickname:  "nickname",
 }
 
+var NicknameListingTableColumns = struct {
+	ID        string
+	CreatedAt string
+	UpdatedAt string
+	DeletedAt string
+	UserID    string
+	GuildID   string
+	Nickname  string
+}{
+	ID:        "nickname_listings.id",
+	CreatedAt: "nickname_listings.created_at",
+	UpdatedAt: "nickname_listings.updated_at",
+	DeletedAt: "nickname_listings.deleted_at",
+	UserID:    "nickname_listings.user_id",
+	GuildID:   "nickname_listings.guild_id",
+	Nickname:  "nickname_listings.nickname",
+}
+
 // Generated where
 
 type whereHelpernull_Int64 struct{ field string }
@@ -64,8 +82,6 @@ func (w whereHelpernull_Int64) EQ(x null.Int64) qm.QueryMod {
 func (w whereHelpernull_Int64) NEQ(x null.Int64) qm.QueryMod {
 	return qmhelper.WhereNullEQ(w.field, true, x)
 }
-func (w whereHelpernull_Int64) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
-func (w whereHelpernull_Int64) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 func (w whereHelpernull_Int64) LT(x null.Int64) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.LT, x)
 }
@@ -78,6 +94,23 @@ func (w whereHelpernull_Int64) GT(x null.Int64) qm.QueryMod {
 func (w whereHelpernull_Int64) GTE(x null.Int64) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
+func (w whereHelpernull_Int64) IN(slice []int64) qm.QueryMod {
+	values := make([]interface{}, 0, len(slice))
+	for _, value := range slice {
+		values = append(values, value)
+	}
+	return qm.WhereIn(fmt.Sprintf("%s IN ?", w.field), values...)
+}
+func (w whereHelpernull_Int64) NIN(slice []int64) qm.QueryMod {
+	values := make([]interface{}, 0, len(slice))
+	for _, value := range slice {
+		values = append(values, value)
+	}
+	return qm.WhereNotIn(fmt.Sprintf("%s NOT IN ?", w.field), values...)
+}
+
+func (w whereHelpernull_Int64) IsNull() qm.QueryMod    { return qmhelper.WhereIsNull(w.field) }
+func (w whereHelpernull_Int64) IsNotNull() qm.QueryMod { return qmhelper.WhereIsNotNull(w.field) }
 
 var NicknameListingWhere = struct {
 	ID        whereHelperint
@@ -115,14 +148,15 @@ type nicknameListingL struct{}
 
 var (
 	nicknameListingAllColumns            = []string{"id", "created_at", "updated_at", "deleted_at", "user_id", "guild_id", "nickname"}
-	nicknameListingColumnsWithoutDefault = []string{"created_at", "updated_at", "deleted_at", "user_id", "guild_id", "nickname"}
-	nicknameListingColumnsWithDefault    = []string{"id"}
+	nicknameListingColumnsWithoutDefault = []string{}
+	nicknameListingColumnsWithDefault    = []string{"id", "created_at", "updated_at", "deleted_at", "user_id", "guild_id", "nickname"}
 	nicknameListingPrimaryKeyColumns     = []string{"id"}
+	nicknameListingGeneratedColumns      = []string{}
 )
 
 type (
 	// NicknameListingSlice is an alias for a slice of pointers to NicknameListing.
-	// This should generally be used opposed to []NicknameListing.
+	// This should almost always be used instead of []NicknameListing.
 	NicknameListingSlice []*NicknameListing
 
 	nicknameListingQuery struct {
@@ -164,7 +198,7 @@ func (q nicknameListingQuery) One(ctx context.Context, exec boil.ContextExecutor
 
 	err := q.Bind(ctx, exec, o)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: failed to execute a one query for nickname_listings")
@@ -190,7 +224,7 @@ func (q nicknameListingQuery) All(ctx context.Context, exec boil.ContextExecutor
 	return o, nil
 }
 
-// CountG returns the count of all NicknameListing records in the query, and panics on error.
+// CountG returns the count of all NicknameListing records in the query using the global executor
 func (q nicknameListingQuery) CountG(ctx context.Context) (int64, error) {
 	return q.Count(ctx, boil.GetContextDB())
 }
@@ -210,7 +244,7 @@ func (q nicknameListingQuery) Count(ctx context.Context, exec boil.ContextExecut
 	return count, nil
 }
 
-// ExistsG checks if the row exists in the table, and panics on error.
+// ExistsG checks if the row exists in the table using the global executor.
 func (q nicknameListingQuery) ExistsG(ctx context.Context) (bool, error) {
 	return q.Exists(ctx, boil.GetContextDB())
 }
@@ -234,7 +268,12 @@ func (q nicknameListingQuery) Exists(ctx context.Context, exec boil.ContextExecu
 // NicknameListings retrieves all the records using an executor.
 func NicknameListings(mods ...qm.QueryMod) nicknameListingQuery {
 	mods = append(mods, qm.From("\"nickname_listings\""))
-	return nicknameListingQuery{NewQuery(mods...)}
+	q := NewQuery(mods...)
+	if len(queries.GetSelect(q)) == 0 {
+		queries.SetSelect(q, []string{"\"nickname_listings\".*"})
+	}
+
+	return nicknameListingQuery{q}
 }
 
 // FindNicknameListingG retrieves a single record by ID.
@@ -259,7 +298,7 @@ func FindNicknameListing(ctx context.Context, exec boil.ContextExecutor, iD int,
 
 	err := q.Bind(ctx, exec, nicknameListingObj)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: unable to select from nickname_listings")
@@ -506,13 +545,13 @@ func (o NicknameListingSlice) UpdateAll(ctx context.Context, exec boil.ContextEx
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *NicknameListing) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
+func (o *NicknameListing) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *NicknameListing) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+func (o *NicknameListing) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
 	if o == nil {
 		return errors.New("models: no nickname_listings provided for upsert")
 	}
@@ -562,12 +601,13 @@ func (o *NicknameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 	var err error
 
 	if !cached {
-		insert, ret := insertColumns.InsertColumnSet(
+		insert, _ := insertColumns.InsertColumnSet(
 			nicknameListingAllColumns,
 			nicknameListingColumnsWithDefault,
 			nicknameListingColumnsWithoutDefault,
 			nzDefaults,
 		)
+
 		update := updateColumns.UpdateColumnSet(
 			nicknameListingAllColumns,
 			nicknameListingPrimaryKeyColumns,
@@ -577,12 +617,18 @@ func (o *NicknameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 			return errors.New("models: unable to upsert nickname_listings, could not build update column list")
 		}
 
+		ret := strmangle.SetComplement(nicknameListingAllColumns, strmangle.SetIntersect(insert, update))
+
 		conflict := conflictColumns
-		if len(conflict) == 0 {
+		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
+			if len(nicknameListingPrimaryKeyColumns) == 0 {
+				return errors.New("models: unable to upsert nickname_listings, could not build conflict column list")
+			}
+
 			conflict = make([]string, len(nicknameListingPrimaryKeyColumns))
 			copy(conflict, nicknameListingPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"nickname_listings\"", updateOnConflict, ret, update, conflict, insert)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"nickname_listings\"", updateOnConflict, ret, update, conflict, insert, opts...)
 
 		cache.valueMapping, err = queries.BindMapping(nicknameListingType, nicknameListingMapping, insert)
 		if err != nil {
@@ -610,7 +656,7 @@ func (o *NicknameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 	}
 	if len(cache.retMapping) != 0 {
 		err = exec.QueryRowContext(ctx, cache.query, vals...).Scan(returns...)
-		if err == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			err = nil // Postgres doesn't return anything when there's no update
 		}
 	} else {
@@ -809,4 +855,9 @@ func NicknameListingExists(ctx context.Context, exec boil.ContextExecutor, iD in
 	}
 
 	return exists, nil
+}
+
+// Exists checks if the NicknameListing row exists.
+func (o *NicknameListing) Exists(ctx context.Context, exec boil.ContextExecutor) (bool, error) {
+	return NicknameListingExists(ctx, exec, o.ID)
 }
