@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.5.0 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -51,6 +51,22 @@ var UsernameListingColumns = struct {
 	Username:  "username",
 }
 
+var UsernameListingTableColumns = struct {
+	ID        string
+	CreatedAt string
+	UpdatedAt string
+	DeletedAt string
+	UserID    string
+	Username  string
+}{
+	ID:        "username_listings.id",
+	CreatedAt: "username_listings.created_at",
+	UpdatedAt: "username_listings.updated_at",
+	DeletedAt: "username_listings.deleted_at",
+	UserID:    "username_listings.user_id",
+	Username:  "username_listings.username",
+}
+
 // Generated where
 
 var UsernameListingWhere = struct {
@@ -87,14 +103,15 @@ type usernameListingL struct{}
 
 var (
 	usernameListingAllColumns            = []string{"id", "created_at", "updated_at", "deleted_at", "user_id", "username"}
-	usernameListingColumnsWithoutDefault = []string{"created_at", "updated_at", "deleted_at", "user_id", "username"}
-	usernameListingColumnsWithDefault    = []string{"id"}
+	usernameListingColumnsWithoutDefault = []string{}
+	usernameListingColumnsWithDefault    = []string{"id", "created_at", "updated_at", "deleted_at", "user_id", "username"}
 	usernameListingPrimaryKeyColumns     = []string{"id"}
+	usernameListingGeneratedColumns      = []string{}
 )
 
 type (
 	// UsernameListingSlice is an alias for a slice of pointers to UsernameListing.
-	// This should generally be used opposed to []UsernameListing.
+	// This should almost always be used instead of []UsernameListing.
 	UsernameListingSlice []*UsernameListing
 
 	usernameListingQuery struct {
@@ -136,7 +153,7 @@ func (q usernameListingQuery) One(ctx context.Context, exec boil.ContextExecutor
 
 	err := q.Bind(ctx, exec, o)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: failed to execute a one query for username_listings")
@@ -162,7 +179,7 @@ func (q usernameListingQuery) All(ctx context.Context, exec boil.ContextExecutor
 	return o, nil
 }
 
-// CountG returns the count of all UsernameListing records in the query, and panics on error.
+// CountG returns the count of all UsernameListing records in the query using the global executor
 func (q usernameListingQuery) CountG(ctx context.Context) (int64, error) {
 	return q.Count(ctx, boil.GetContextDB())
 }
@@ -182,7 +199,7 @@ func (q usernameListingQuery) Count(ctx context.Context, exec boil.ContextExecut
 	return count, nil
 }
 
-// ExistsG checks if the row exists in the table, and panics on error.
+// ExistsG checks if the row exists in the table using the global executor.
 func (q usernameListingQuery) ExistsG(ctx context.Context) (bool, error) {
 	return q.Exists(ctx, boil.GetContextDB())
 }
@@ -206,7 +223,12 @@ func (q usernameListingQuery) Exists(ctx context.Context, exec boil.ContextExecu
 // UsernameListings retrieves all the records using an executor.
 func UsernameListings(mods ...qm.QueryMod) usernameListingQuery {
 	mods = append(mods, qm.From("\"username_listings\""))
-	return usernameListingQuery{NewQuery(mods...)}
+	q := NewQuery(mods...)
+	if len(queries.GetSelect(q)) == 0 {
+		queries.SetSelect(q, []string{"\"username_listings\".*"})
+	}
+
+	return usernameListingQuery{q}
 }
 
 // FindUsernameListingG retrieves a single record by ID.
@@ -231,7 +253,7 @@ func FindUsernameListing(ctx context.Context, exec boil.ContextExecutor, iD int,
 
 	err := q.Bind(ctx, exec, usernameListingObj)
 	if err != nil {
-		if errors.Cause(err) == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, sql.ErrNoRows
 		}
 		return nil, errors.Wrap(err, "models: unable to select from username_listings")
@@ -478,13 +500,13 @@ func (o UsernameListingSlice) UpdateAll(ctx context.Context, exec boil.ContextEx
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *UsernameListing) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
+func (o *UsernameListing) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *UsernameListing) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+func (o *UsernameListing) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
 	if o == nil {
 		return errors.New("models: no username_listings provided for upsert")
 	}
@@ -534,12 +556,13 @@ func (o *UsernameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 	var err error
 
 	if !cached {
-		insert, ret := insertColumns.InsertColumnSet(
+		insert, _ := insertColumns.InsertColumnSet(
 			usernameListingAllColumns,
 			usernameListingColumnsWithDefault,
 			usernameListingColumnsWithoutDefault,
 			nzDefaults,
 		)
+
 		update := updateColumns.UpdateColumnSet(
 			usernameListingAllColumns,
 			usernameListingPrimaryKeyColumns,
@@ -549,12 +572,18 @@ func (o *UsernameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 			return errors.New("models: unable to upsert username_listings, could not build update column list")
 		}
 
+		ret := strmangle.SetComplement(usernameListingAllColumns, strmangle.SetIntersect(insert, update))
+
 		conflict := conflictColumns
-		if len(conflict) == 0 {
+		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
+			if len(usernameListingPrimaryKeyColumns) == 0 {
+				return errors.New("models: unable to upsert username_listings, could not build conflict column list")
+			}
+
 			conflict = make([]string, len(usernameListingPrimaryKeyColumns))
 			copy(conflict, usernameListingPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"username_listings\"", updateOnConflict, ret, update, conflict, insert)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"username_listings\"", updateOnConflict, ret, update, conflict, insert, opts...)
 
 		cache.valueMapping, err = queries.BindMapping(usernameListingType, usernameListingMapping, insert)
 		if err != nil {
@@ -582,7 +611,7 @@ func (o *UsernameListing) Upsert(ctx context.Context, exec boil.ContextExecutor,
 	}
 	if len(cache.retMapping) != 0 {
 		err = exec.QueryRowContext(ctx, cache.query, vals...).Scan(returns...)
-		if err == sql.ErrNoRows {
+		if errors.Is(err, sql.ErrNoRows) {
 			err = nil // Postgres doesn't return anything when there's no update
 		}
 	} else {
@@ -781,4 +810,9 @@ func UsernameListingExists(ctx context.Context, exec boil.ContextExecutor, iD in
 	}
 
 	return exists, nil
+}
+
+// Exists checks if the UsernameListing row exists.
+func (o *UsernameListing) Exists(ctx context.Context, exec boil.ContextExecutor) (bool, error) {
+	return UsernameListingExists(ctx, exec, o.ID)
 }
