@@ -54,11 +54,18 @@ var CmdActivateTestEntitlement = &commands.YAGCommand{
 	RunInDM:              true,
 	Arguments: []*dcmd.ArgDef{
 		{Name: "User", Type: dcmd.UserID},
+		{Name: "SKUID", Type: dcmd.Int},
 	},
 	RunFunc: util.RequireOwner(func(data *dcmd.Data) (interface{}, error) {
 		userID := data.Args[0].Int64()
+		skuID := data.Args[1].Int64()
+		slots := fetchSlotsForDiscordSku(int64(skuID))
+		if slots == 0 {
+			return fmt.Sprintf("No Slots configured for skuID %d", skuID), nil
+		}
+
 		entitlementData := &discordgo.EntitlementTest{
-			SKUID:     int64(confDiscordPremiumSKUID.GetInt()),
+			SKUID:     skuID,
 			OwnerID:   userID,
 			OwnerType: discordgo.EntitlementOwnerTypeUserSubscription,
 		}
@@ -66,7 +73,7 @@ var CmdActivateTestEntitlement = &commands.YAGCommand{
 		if err != nil {
 			return fmt.Sprintf("Failed enabling Entitlement for <@%d>: %s", userID, err), nil
 		}
-		return fmt.Sprintf("Enabled Entitlement for <@%d>", userID), nil
+		return fmt.Sprintf("Enabled Entitlement with %d slots for <@%d>", slots, userID), nil
 	}),
 }
 
@@ -80,12 +87,18 @@ var CmdDeleteTestEntitlement = &commands.YAGCommand{
 	RunInDM:              true,
 	Arguments: []*dcmd.ArgDef{
 		{Name: "User", Type: dcmd.UserID},
+		{Name: "SKUID", Type: dcmd.Int},
 	},
 	RunFunc: util.RequireOwner(func(data *dcmd.Data) (interface{}, error) {
 		userID := data.Args[0].Int64()
+		skuID := data.Args[1].Int64()
+		slots := fetchSlotsForDiscordSku(int64(skuID))
+		if slots == 0 {
+			return fmt.Sprintf("No Slots configured for skuID %d", skuID), nil
+		}
 		entitlements, err := common.BotSession.Entitlements(common.BotApplication.ID, &discordgo.EntitlementFilterOptions{
 			UserID:       userID,
-			SkuIDs:       []int64{int64(confDiscordPremiumSKUID.GetInt())},
+			SkuIDs:       []int64{skuID},
 			ExcludeEnded: true,
 		})
 		if err != nil {
@@ -95,12 +108,10 @@ var CmdDeleteTestEntitlement = &commands.YAGCommand{
 			return fmt.Sprintf("No Entitlements found for <@%d>", userID), nil
 		}
 		for _, v := range entitlements {
-			//if v.Type == discordgo.EntitlementTypeTestModePurchase {
 			err := common.BotSession.EntitlementTestDelete(common.BotApplication.ID, v.ID)
 			if err != nil {
 				return fmt.Sprintf("Failed deleting Entitlement for <@%d>: %s", userID, err), nil
 			}
-			//}
 		}
 		return fmt.Sprintf("Deleted Entitlement for <@%d>", userID), nil
 	}),
