@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.16.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -44,6 +44,7 @@ type ReputationConfig struct {
 	RequiredReceiveRoles      types.Int64Array `boil:"required_receive_roles" json:"required_receive_roles,omitempty" toml:"required_receive_roles" yaml:"required_receive_roles,omitempty"`
 	BlacklistedGiveRoles      types.Int64Array `boil:"blacklisted_give_roles" json:"blacklisted_give_roles,omitempty" toml:"blacklisted_give_roles" yaml:"blacklisted_give_roles,omitempty"`
 	BlacklistedReceiveRoles   types.Int64Array `boil:"blacklisted_receive_roles" json:"blacklisted_receive_roles,omitempty" toml:"blacklisted_receive_roles" yaml:"blacklisted_receive_roles,omitempty"`
+	ThanksRegex               null.String      `boil:"thanks_regex" json:"thanks_regex,omitempty" toml:"thanks_regex" yaml:"thanks_regex,omitempty"`
 
 	R *reputationConfigR `boil:"-" json:"-" toml:"-" yaml:"-"`
 	L reputationConfigL  `boil:"-" json:"-" toml:"-" yaml:"-"`
@@ -69,6 +70,7 @@ var ReputationConfigColumns = struct {
 	RequiredReceiveRoles      string
 	BlacklistedGiveRoles      string
 	BlacklistedReceiveRoles   string
+	ThanksRegex               string
 }{
 	GuildID:                   "guild_id",
 	PointsName:                "points_name",
@@ -89,6 +91,7 @@ var ReputationConfigColumns = struct {
 	RequiredReceiveRoles:      "required_receive_roles",
 	BlacklistedGiveRoles:      "blacklisted_give_roles",
 	BlacklistedReceiveRoles:   "blacklisted_receive_roles",
+	ThanksRegex:               "thanks_regex",
 }
 
 var ReputationConfigTableColumns = struct {
@@ -111,6 +114,7 @@ var ReputationConfigTableColumns = struct {
 	RequiredReceiveRoles      string
 	BlacklistedGiveRoles      string
 	BlacklistedReceiveRoles   string
+	ThanksRegex               string
 }{
 	GuildID:                   "reputation_configs.guild_id",
 	PointsName:                "reputation_configs.points_name",
@@ -131,6 +135,7 @@ var ReputationConfigTableColumns = struct {
 	RequiredReceiveRoles:      "reputation_configs.required_receive_roles",
 	BlacklistedGiveRoles:      "reputation_configs.blacklisted_give_roles",
 	BlacklistedReceiveRoles:   "reputation_configs.blacklisted_receive_roles",
+	ThanksRegex:               "reputation_configs.thanks_regex",
 }
 
 // Generated where
@@ -160,16 +165,12 @@ func (w whereHelperint64) NIN(slice []int64) qm.QueryMod {
 
 type whereHelperstring struct{ field string }
 
-func (w whereHelperstring) EQ(x string) qm.QueryMod     { return qmhelper.Where(w.field, qmhelper.EQ, x) }
-func (w whereHelperstring) NEQ(x string) qm.QueryMod    { return qmhelper.Where(w.field, qmhelper.NEQ, x) }
-func (w whereHelperstring) LT(x string) qm.QueryMod     { return qmhelper.Where(w.field, qmhelper.LT, x) }
-func (w whereHelperstring) LTE(x string) qm.QueryMod    { return qmhelper.Where(w.field, qmhelper.LTE, x) }
-func (w whereHelperstring) GT(x string) qm.QueryMod     { return qmhelper.Where(w.field, qmhelper.GT, x) }
-func (w whereHelperstring) GTE(x string) qm.QueryMod    { return qmhelper.Where(w.field, qmhelper.GTE, x) }
-func (w whereHelperstring) LIKE(x string) qm.QueryMod   { return qm.Where(w.field+" LIKE ?", x) }
-func (w whereHelperstring) NLIKE(x string) qm.QueryMod  { return qm.Where(w.field+" NOT LIKE ?", x) }
-func (w whereHelperstring) ILIKE(x string) qm.QueryMod  { return qm.Where(w.field+" ILIKE ?", x) }
-func (w whereHelperstring) NILIKE(x string) qm.QueryMod { return qm.Where(w.field+" NOT ILIKE ?", x) }
+func (w whereHelperstring) EQ(x string) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.EQ, x) }
+func (w whereHelperstring) NEQ(x string) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.NEQ, x) }
+func (w whereHelperstring) LT(x string) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.LT, x) }
+func (w whereHelperstring) LTE(x string) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.LTE, x) }
+func (w whereHelperstring) GT(x string) qm.QueryMod  { return qmhelper.Where(w.field, qmhelper.GT, x) }
+func (w whereHelperstring) GTE(x string) qm.QueryMod { return qmhelper.Where(w.field, qmhelper.GTE, x) }
 func (w whereHelperstring) IN(slice []string) qm.QueryMod {
 	values := make([]interface{}, 0, len(slice))
 	for _, value := range slice {
@@ -237,18 +238,6 @@ func (w whereHelpernull_String) GT(x null.String) qm.QueryMod {
 func (w whereHelpernull_String) GTE(x null.String) qm.QueryMod {
 	return qmhelper.Where(w.field, qmhelper.GTE, x)
 }
-func (w whereHelpernull_String) LIKE(x null.String) qm.QueryMod {
-	return qm.Where(w.field+" LIKE ?", x)
-}
-func (w whereHelpernull_String) NLIKE(x null.String) qm.QueryMod {
-	return qm.Where(w.field+" NOT LIKE ?", x)
-}
-func (w whereHelpernull_String) ILIKE(x null.String) qm.QueryMod {
-	return qm.Where(w.field+" ILIKE ?", x)
-}
-func (w whereHelpernull_String) NILIKE(x null.String) qm.QueryMod {
-	return qm.Where(w.field+" NOT ILIKE ?", x)
-}
 func (w whereHelpernull_String) IN(slice []string) qm.QueryMod {
 	values := make([]interface{}, 0, len(slice))
 	for _, value := range slice {
@@ -311,6 +300,7 @@ var ReputationConfigWhere = struct {
 	RequiredReceiveRoles      whereHelpertypes_Int64Array
 	BlacklistedGiveRoles      whereHelpertypes_Int64Array
 	BlacklistedReceiveRoles   whereHelpertypes_Int64Array
+	ThanksRegex               whereHelpernull_String
 }{
 	GuildID:                   whereHelperint64{field: "\"reputation_configs\".\"guild_id\""},
 	PointsName:                whereHelperstring{field: "\"reputation_configs\".\"points_name\""},
@@ -331,6 +321,7 @@ var ReputationConfigWhere = struct {
 	RequiredReceiveRoles:      whereHelpertypes_Int64Array{field: "\"reputation_configs\".\"required_receive_roles\""},
 	BlacklistedGiveRoles:      whereHelpertypes_Int64Array{field: "\"reputation_configs\".\"blacklisted_give_roles\""},
 	BlacklistedReceiveRoles:   whereHelpertypes_Int64Array{field: "\"reputation_configs\".\"blacklisted_receive_roles\""},
+	ThanksRegex:               whereHelpernull_String{field: "\"reputation_configs\".\"thanks_regex\""},
 }
 
 // ReputationConfigRels is where relationship names are stored.
@@ -350,9 +341,9 @@ func (*reputationConfigR) NewStruct() *reputationConfigR {
 type reputationConfigL struct{}
 
 var (
-	reputationConfigAllColumns            = []string{"guild_id", "points_name", "enabled", "cooldown", "max_give_amount", "required_give_role", "required_receive_role", "blacklisted_give_role", "blacklisted_receive_role", "admin_role", "disable_thanks_detection", "whitelisted_thanks_channels", "blacklisted_thanks_channels", "max_remove_amount", "admin_roles", "required_give_roles", "required_receive_roles", "blacklisted_give_roles", "blacklisted_receive_roles"}
+	reputationConfigAllColumns            = []string{"guild_id", "points_name", "enabled", "cooldown", "max_give_amount", "required_give_role", "required_receive_role", "blacklisted_give_role", "blacklisted_receive_role", "admin_role", "disable_thanks_detection", "whitelisted_thanks_channels", "blacklisted_thanks_channels", "max_remove_amount", "admin_roles", "required_give_roles", "required_receive_roles", "blacklisted_give_roles", "blacklisted_receive_roles", "thanks_regex"}
 	reputationConfigColumnsWithoutDefault = []string{"guild_id", "points_name", "enabled", "cooldown", "max_give_amount"}
-	reputationConfigColumnsWithDefault    = []string{"required_give_role", "required_receive_role", "blacklisted_give_role", "blacklisted_receive_role", "admin_role", "disable_thanks_detection", "whitelisted_thanks_channels", "blacklisted_thanks_channels", "max_remove_amount", "admin_roles", "required_give_roles", "required_receive_roles", "blacklisted_give_roles", "blacklisted_receive_roles"}
+	reputationConfigColumnsWithDefault    = []string{"required_give_role", "required_receive_role", "blacklisted_give_role", "blacklisted_receive_role", "admin_role", "disable_thanks_detection", "whitelisted_thanks_channels", "blacklisted_thanks_channels", "max_remove_amount", "admin_roles", "required_give_roles", "required_receive_roles", "blacklisted_give_roles", "blacklisted_receive_roles", "thanks_regex"}
 	reputationConfigPrimaryKeyColumns     = []string{"guild_id"}
 	reputationConfigGeneratedColumns      = []string{}
 )
@@ -732,13 +723,13 @@ func (o ReputationConfigSlice) UpdateAll(ctx context.Context, exec boil.ContextE
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *ReputationConfig) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
+func (o *ReputationConfig) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *ReputationConfig) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+func (o *ReputationConfig) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
 	if o == nil {
 		return errors.New("models: no reputation_configs provided for upsert")
 	}
@@ -780,7 +771,7 @@ func (o *ReputationConfig) Upsert(ctx context.Context, exec boil.ContextExecutor
 	var err error
 
 	if !cached {
-		insert, _ := insertColumns.InsertColumnSet(
+		insert, ret := insertColumns.InsertColumnSet(
 			reputationConfigAllColumns,
 			reputationConfigColumnsWithDefault,
 			reputationConfigColumnsWithoutDefault,
@@ -796,18 +787,12 @@ func (o *ReputationConfig) Upsert(ctx context.Context, exec boil.ContextExecutor
 			return errors.New("models: unable to upsert reputation_configs, could not build update column list")
 		}
 
-		ret := strmangle.SetComplement(reputationConfigAllColumns, strmangle.SetIntersect(insert, update))
-
 		conflict := conflictColumns
-		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
-			if len(reputationConfigPrimaryKeyColumns) == 0 {
-				return errors.New("models: unable to upsert reputation_configs, could not build conflict column list")
-			}
-
+		if len(conflict) == 0 {
 			conflict = make([]string, len(reputationConfigPrimaryKeyColumns))
 			copy(conflict, reputationConfigPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_configs\"", updateOnConflict, ret, update, conflict, insert, opts...)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_configs\"", updateOnConflict, ret, update, conflict, insert)
 
 		cache.valueMapping, err = queries.BindMapping(reputationConfigType, reputationConfigMapping, insert)
 		if err != nil {
