@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.16.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -532,13 +532,13 @@ func (o ReputationLogSlice) UpdateAll(ctx context.Context, exec boil.ContextExec
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *ReputationLog) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
+func (o *ReputationLog) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *ReputationLog) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+func (o *ReputationLog) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
 	if o == nil {
 		return errors.New("models: no reputation_log provided for upsert")
 	}
@@ -587,7 +587,7 @@ func (o *ReputationLog) Upsert(ctx context.Context, exec boil.ContextExecutor, u
 	var err error
 
 	if !cached {
-		insert, _ := insertColumns.InsertColumnSet(
+		insert, ret := insertColumns.InsertColumnSet(
 			reputationLogAllColumns,
 			reputationLogColumnsWithDefault,
 			reputationLogColumnsWithoutDefault,
@@ -603,18 +603,12 @@ func (o *ReputationLog) Upsert(ctx context.Context, exec boil.ContextExecutor, u
 			return errors.New("models: unable to upsert reputation_log, could not build update column list")
 		}
 
-		ret := strmangle.SetComplement(reputationLogAllColumns, strmangle.SetIntersect(insert, update))
-
 		conflict := conflictColumns
-		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
-			if len(reputationLogPrimaryKeyColumns) == 0 {
-				return errors.New("models: unable to upsert reputation_log, could not build conflict column list")
-			}
-
+		if len(conflict) == 0 {
 			conflict = make([]string, len(reputationLogPrimaryKeyColumns))
 			copy(conflict, reputationLogPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_log\"", updateOnConflict, ret, update, conflict, insert, opts...)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_log\"", updateOnConflict, ret, update, conflict, insert)
 
 		cache.valueMapping, err = queries.BindMapping(reputationLogType, reputationLogMapping, insert)
 		if err != nil {
