@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.16.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -469,13 +469,13 @@ func (o ReputationRoleSlice) UpdateAll(ctx context.Context, exec boil.ContextExe
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *ReputationRole) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
+func (o *ReputationRole) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *ReputationRole) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+func (o *ReputationRole) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
 	if o == nil {
 		return errors.New("models: no reputation_roles provided for upsert")
 	}
@@ -517,7 +517,7 @@ func (o *ReputationRole) Upsert(ctx context.Context, exec boil.ContextExecutor, 
 	var err error
 
 	if !cached {
-		insert, _ := insertColumns.InsertColumnSet(
+		insert, ret := insertColumns.InsertColumnSet(
 			reputationRoleAllColumns,
 			reputationRoleColumnsWithDefault,
 			reputationRoleColumnsWithoutDefault,
@@ -533,18 +533,12 @@ func (o *ReputationRole) Upsert(ctx context.Context, exec boil.ContextExecutor, 
 			return errors.New("models: unable to upsert reputation_roles, could not build update column list")
 		}
 
-		ret := strmangle.SetComplement(reputationRoleAllColumns, strmangle.SetIntersect(insert, update))
-
 		conflict := conflictColumns
-		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
-			if len(reputationRolePrimaryKeyColumns) == 0 {
-				return errors.New("models: unable to upsert reputation_roles, could not build conflict column list")
-			}
-
+		if len(conflict) == 0 {
 			conflict = make([]string, len(reputationRolePrimaryKeyColumns))
 			copy(conflict, reputationRolePrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_roles\"", updateOnConflict, ret, update, conflict, insert, opts...)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_roles\"", updateOnConflict, ret, update, conflict, insert)
 
 		cache.valueMapping, err = queries.BindMapping(reputationRoleType, reputationRoleMapping, insert)
 		if err != nil {
