@@ -1,4 +1,4 @@
-// Code generated by SQLBoiler 4.16.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
+// Code generated by SQLBoiler 4.14.2 (https://github.com/volatiletech/sqlboiler). DO NOT EDIT.
 // This file is meant to be re-generated in place and/or deleted at any time.
 
 package models
@@ -476,13 +476,13 @@ func (o ReputationUserSlice) UpdateAll(ctx context.Context, exec boil.ContextExe
 }
 
 // UpsertG attempts an insert, and does an update or ignore on conflict.
-func (o *ReputationUser) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
-	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns, opts...)
+func (o *ReputationUser) UpsertG(ctx context.Context, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
+	return o.Upsert(ctx, boil.GetContextDB(), updateOnConflict, conflictColumns, updateColumns, insertColumns)
 }
 
 // Upsert attempts an insert using an executor, and does an update or ignore on conflict.
 // See boil.Columns documentation for how to properly use updateColumns and insertColumns.
-func (o *ReputationUser) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns, opts ...UpsertOptionFunc) error {
+func (o *ReputationUser) Upsert(ctx context.Context, exec boil.ContextExecutor, updateOnConflict bool, conflictColumns []string, updateColumns, insertColumns boil.Columns) error {
 	if o == nil {
 		return errors.New("models: no reputation_users provided for upsert")
 	}
@@ -531,7 +531,7 @@ func (o *ReputationUser) Upsert(ctx context.Context, exec boil.ContextExecutor, 
 	var err error
 
 	if !cached {
-		insert, _ := insertColumns.InsertColumnSet(
+		insert, ret := insertColumns.InsertColumnSet(
 			reputationUserAllColumns,
 			reputationUserColumnsWithDefault,
 			reputationUserColumnsWithoutDefault,
@@ -547,18 +547,12 @@ func (o *ReputationUser) Upsert(ctx context.Context, exec boil.ContextExecutor, 
 			return errors.New("models: unable to upsert reputation_users, could not build update column list")
 		}
 
-		ret := strmangle.SetComplement(reputationUserAllColumns, strmangle.SetIntersect(insert, update))
-
 		conflict := conflictColumns
-		if len(conflict) == 0 && updateOnConflict && len(update) != 0 {
-			if len(reputationUserPrimaryKeyColumns) == 0 {
-				return errors.New("models: unable to upsert reputation_users, could not build conflict column list")
-			}
-
+		if len(conflict) == 0 {
 			conflict = make([]string, len(reputationUserPrimaryKeyColumns))
 			copy(conflict, reputationUserPrimaryKeyColumns)
 		}
-		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_users\"", updateOnConflict, ret, update, conflict, insert, opts...)
+		cache.query = buildUpsertQueryPostgres(dialect, "\"reputation_users\"", updateOnConflict, ret, update, conflict, insert)
 
 		cache.valueMapping, err = queries.BindMapping(reputationUserType, reputationUserMapping, insert)
 		if err != nil {
